% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_Citations.R
\docType{data}
\name{Author_Citations}
\alias{Author_Citations}
\title{Author_Citations.}
\format{
\subsection{\code{Author_Citations}}{

Data frame that contains.
\describe{
\item{Name}{Author name.}
\item{Citations per article}{How many citations on average does the author get on his articles.}
\item{Articles}{How many articles the author has published.}
\item{Citations}{How many times the published article has been cited.}
\item{Rank}{The ranking by how many total citations the author has gotten.}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Author_Citations
}
\description{
How many times a given author was referenced in ISI hournals from 1981 to 1997. Authors with the same name are merged.
}
\details{
" 'This data covers all publications from ISI catalogued journals
from January 1981 - June 1997 and was compiled by H. Small and D. Pendlebury
of the Institute for Scientific Information. This file includes several
corrections to the data originally supplied by the ISI: missing decimal
points in several entries are now provided, and the data is correctly
ordered according to total number of citations (a block of the original
data that was provided by the ISI was not in correct rank order). Notice also
that different authors with the same last name and first initial are erroneously
lumped into one entry in the table (for example #12 or #397 on this list).
Use this list with caution! ' Quoted from:
Citations of the 1120 Most-Cited Physicists 1981 - June 1997. (n.d.).
Retrieved September 11, 2024, from
https://physics.bu.edu/~redner/projects/citation/physics-by-person.html"
}
\keyword{datasets}
