% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_compare.R
\name{BC_compare}
\alias{BC_compare}
\title{Compare the linear and the nls methods parameters}
\usage{
BC_compare(
  df_abundance = NULL,
  column = NULL,
  BC_plot_list = NULL,
  c_gfx_title = "Linear vs nls model error",
  c_gfx_label = TRUE,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_plot_list}{A list that contains 2 objects previously generated with \code{BC_plot}. The first one must use the linear paramenters and the second one parameters estimated by the nls method.}

\item{c_gfx_title}{String. Changes the title of the graph.}

\item{c_gfx_label}{Logical. Adds a label that adds the model_extra data of both models. Defaults to true.}

\item{...}{passes arguments to \code{BC_plot}.}
}
\value{
A list with that includes a graph and a data frame with difference data between predicted and real values per model.
}
\description{
Creates a graph that illustrates error in the predictions of the linear and nls models for quick comparison. The y axis is the difference between the predicted value and the original abundance value. This value is then made absolute and 1 is added before taking its logarithm. The addition of 1 keeps a difference of 0 as 0 in the plot and prevents a small decimal linear difference from becoming a large difference in the plot. The x axis is the rank for that given error. This allows us to see where in the distribution each model performs better or worse.
}
\examples{
comparehmp_wgs <- BC_compare(hmp_wgs,2)
comparehmp_wgs[[1]]
head(comparehmp_wgs[[2]])

compareEC_Metabolite <- BC_compare(EC_Metabolite, column = 2,model_extra="S")
compareEC_Metabolite[[1]]
head(compareEC_Metabolite[[2]])
}
