% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_param.R
\name{BC_param}
\alias{BC_param}
\title{Estimate the DGB Distribution Parameters}
\usage{
BC_param(
  df_abundance = NULL,
  column_param = NULL,
  confidence_interval = 0.95,
  BC_rank_object = NULL,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column_param}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{confidence_interval}{The confidence interval to calculate for the DGB distribution.}

\item{BC_rank_object}{Optional. A previous object generated by \code{BC_rank}.}

\item{...}{passes arguments to \code{BC_rank}.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data and predicted confidence interval values, the adjusted parameters, the confidence interval of the parameters, the linear model and a summary of the model.
}
\description{
Estimates the parameters of the DGB distribution proposed in Martinez-Mekler et al. (2009) \doi{10.1371/journal.pone.0004791} for a given set of data.
\code{BC_param} calculates the log of the data and estimates the abundance from the ranking using a linear model.
The coefficients of the linear model are then scaled for posterior use.
}
\examples{
BC_param(hmp_wgs,2,confidence_interval=0.99)

BC_param(Tara_Data,column_param=2,is_phyloseq=TRUE,confidence_interval=0.9)

}
