% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_plot.R
\name{BC_plot}
\alias{BC_plot}
\title{Plot DGB Distribution and Observed Data}
\usage{
BC_plot(
  df_abundance = NULL,
  column_plot = NULL,
  BC_model_object = NULL,
  obs = TRUE,
  obs_shape = 16,
  obs_col = "#78a7ff",
  obs_size = 1,
  model = TRUE,
  model_col = "#000000",
  model_width = 0.5,
  confint = TRUE,
  confint_col = "#ed8666",
  confint_width = 1,
  confrange = TRUE,
  confrange_col = "#ffd078",
  gfx_alpha = 0.75,
  gfx_title = "Rank-Abundance Diagram",
  gfx_label = TRUE,
  plot_silent = FALSE,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column_plot}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_model_object}{Optional. A previous object generated by \code{BC_model}.}

\item{obs}{Logical. Whether to plot the observed abundance data. Defaults to true.}

\item{obs_shape}{Numerical. The shape of the plotted observed abundance data.}

\item{obs_col}{The color for the observations.}

\item{obs_size}{Numeric. The size for the observations.}

\item{model}{Logical. Whether to show the models predicted data. Defaults to true.}

\item{model_col}{Specify a color for the model.}

\item{model_width}{Numeric. Changes the width of the lines to use for the model.}

\item{confint}{Logical. Whether to add the confidence interval lines. Defaults to true.}

\item{confint_col}{Specify a color for the confidence interval lines.}

\item{confint_width}{Numeric. Changes the width of the confidence interval lines.}

\item{confrange}{Logical. Whether to shade the area in the confidence interval. Defaults to true.}

\item{confrange_col}{Specify a color to use for the confidence interval shading.}

\item{gfx_alpha}{Numeric. Modifies all of the graphed objects alpha. Default=0.75.}

\item{gfx_title}{String. Changes the title of the graph.}

\item{gfx_label}{Logical. Whether to show the parameters used and r2.}

\item{plot_silent}{Logical. Whether to print to console the output list and plot the graph.}

\item{...}{passes arguments to \code{BC_model}.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data, model data and confidence interval data, the adjusted parameters, the confidence interval of the parameters, the linear model, a summary of the model, a generated function for use with raw numeric data and a ggplot2 object that shows the DGBD distribution and observed data.
}
\description{
Creates a graphical representation of the DGB Distribution model generated by \code{BC_model}.
}
\examples{
BC_plot(hmp_wgs,2)

BC_plot(hmp_wgs,2,obs=FALSE)

BC_plot(hmp_wgs,2,confint=FALSE,confrange=FALSE,model_width=2)

BC_plot(Tara_Data,column_plot=1,is_phyloseq=TRUE)

}
