% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_rank.R
\name{BC_rank}
\alias{BC_rank}
\title{Rank the Abundance from a File or Data Frame}
\usage{
BC_rank(
  df_abundance,
  column_rank,
  rank_threshold = 0,
  is_phyloseq = FALSE,
  ...
)
}
\arguments{
\item{df_abundance}{The data frame to use. A string that contains a path can be used to load files.}

\item{column_rank}{Which column will be ranked. Can be a string or a number.}

\item{rank_threshold}{Numeric, optional. The abundance values to analyze in the sample have to be higher than this number. Defaults to 0.}

\item{is_phyloseq}{Logical, optional. Use if the object to be analysed has the phyloseq class.}

\item{...}{passes arguments to \code{read.table}.}
}
\value{
The input data frame with a column appended to the left that stores the generated ranking. The column used for ranking is renamed "abundance".
}
\description{
Generates a ranking from a column that stores abundance information in a data frame or file.
}
\examples{
BC_rank(hmp_wgs,2)

BC_rank(Tara_Data,3,is_phyloseq=TRUE)

}
