% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Citation_ISICatalogued.R
\docType{data}
\name{Citation_ISICatalogued}
\alias{Citation_ISICatalogued}
\title{Citation_ISICatalogued.}
\format{
\subsection{\code{Citation_ISICatalogued}}{

Data frame that contains.
\describe{
\item{N}{For a given number of citations:}
\item{Number.of.papers.with.N.citations}{How many articles there are.}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Citation_ISICatalogued
}
\description{
Distribution of how many times a paper was cited, taken from ISI journals.
}
\details{
" 'This distribution covers all publications from ISI catalogued
journals that were published in 1981 and cited during the period January
1981 - June 1997. This information was compiled by H. Small and D. Pendlebury
of the Institute for Scientific Information.' Quoted from:
ISI Citation Data. (n.d.). Retrieved September 11, 2024, from
https://physics.bu.edu/~redner/projects/citation/isi.html  "
}
\keyword{datasets}
