% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_Codon.R
\docType{data}
\name{EC_Codon}
\alias{EC_Codon}
\title{EC_Codon.}
\format{
\subsection{\code{EC_Codon}}{

Data frame that contains.
\describe{
\item{NC_008253}{Codon freq for NC_008253 strain.}
\item{NC_008563}{Codon freq for NC_008563 strain.}
\item{NC_010468}{Codon freq for NC_010468 strain.}
\item{NC_004431}{Codon freq for NC_004431 strain.}
\item{NC_009801}{Codon freq for NC_009801 strain.}
\item{NC_009800}{Codon freq for NC_009800 strain.}
\item{NC_002655}{Codon freq for NC_002655 strain.}
\item{NC_002695}{Codon freq for NC_002695 strain.}
\item{NC_010498}{Codon freq for NC_010498 strain.}
\item{NC_007946}{Codon freq for NC_007946 strain.}
\item{NC_010473}{Codon freq for NC_010473 strain.}
\item{NC_000913}{Codon freq for NC_000913 strain.}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
EC_Codon
}
\description{
The codon frequency of a dozen different strains of Eschericia coli.
}
\details{
"Pending citation"
}
\keyword{datasets}
