% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_Metabolite.R
\docType{data}
\name{EC_Metabolite}
\alias{EC_Metabolite}
\title{EC_Metabolite.}
\format{
\subsection{\code{EC_Metabolite}}{

Data frame that contains.
\describe{
\item{degree}{The number of connections the metabolite has in the network.}
\item{count}{How many metabolites have this number of connections (degree distribution of the network).}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
EC_Metabolite
}
\description{
The degrees of metabolites in the metabolic network of the bacterium Escherichia coli.
}
\details{
" Mirrored at
https://aaronclauset.github.io/powerlaws/data.htm"
}
\keyword{datasets}
