% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flares.R
\docType{data}
\name{Flares}
\alias{Flares}
\title{Flares.}
\format{
\subsection{\code{Flares}}{

Data frame that contains.
\describe{
\item{intensity}{The number of links pointing to the page}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Flares
}
\description{
Peak gamma-ray intensity of solar flares.
}
\details{
" 'The cumulative distribution of the peak
gamma-ray intensity of solar flares. The observations were made between 1980
and 1989 by the instrument known as the Hard X-Ray Burst Spectrometer aboard
the Solar Maximum Mission satellite launched in 1980. The spectrometer used a
CsI scintillation detector to measure gamma-rays from solar flare...
The data are from the NASA Goddard Space Flight Center,
umbra.nascom.nasa.gov/smm/hxrbs.html.' Quoted from:
Newman, M. (2005). Power laws, Pareto distributions and Zipf’s law.
Contemporary Physics, 46(5), 323–351. https://doi.org/10.1080/00107510500052444"
}
\keyword{datasets}
