% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmp_wgs.R
\docType{data}
\name{hmp_wgs}
\alias{hmp_wgs}
\title{Human Microbiome Project WGS .}
\format{
\subsection{\code{hmp_wgs}}{

Data frame that contains.
\describe{
\item{PID}{Module Table ID that identifies organisms that contain complete KEGG modules}
\item{Abundance}{Abundance data}
...
}
}
}
\source{
\url{https://portal.hmpdacc.org/files/1670203039de370df9a35a043735e964}
}
\usage{
hmp_wgs
}
\description{
Whole Genome Sequencing data obtained from the Human Microbiome Project production phase I.
}
\details{
"The overall mission of the HMP is to generate resources to facilitate
characterization of the human microbiota to further our understanding of how the
microbiome impacts human health and disease. We are currently in the second phase
of the HMP. In this phase, the iHMP will create integrated longitudinal datasets
from both the microbiome and host from three different cohort studies of
microbiome-associated conditions using multiple ‘omics technologies. Each of these
study groups has engaged in providing new computational tools and integrative molecular
perspectives on microbial activity during dysbiosis. As a result of creating these
multi’omic data resources, the iHMP has opened up new opportunities for data integration
in the human microbiome.".
\url{https://www.hmpdacc.org/ihmp/}
}
\keyword{datasets}
