% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quakes.R
\docType{data}
\name{Quakes}
\alias{Quakes}
\title{Quakes.}
\format{
\subsection{\code{Quakes}}{

Data frame that contains.
\describe{
\item{magnitude}{The magnitude of an earthquake}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Quakes
}
\description{
The distribution of earthquake magnitudes in California between 1910 and 1992.
}
\details{
" 'Magnitude of earthquakes: The cumulative distribution of
the Richter (local) magnitude of earthquakes occurring in California between
January 1910 and May 1992, as recorded in the Berkeley Earthquake Catalog.
The Richter magnitude is defined as the logarithm, base 10, of the maximum
amplitude of motion detected in the earthquake. ... The data are from the
National Geophysical Data Center, www.ngdc.noaa.gov.' Quoted from:
Newman, M. (2005). Power laws, Pareto distributions and Zipf’s law.
Contemporary Physics, 46(5), 323–351. https://doi.org/10.1080/00107510500052444"
}
\keyword{datasets}
