% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tara_Data.R
\docType{data}
\name{Tara_Data}
\alias{Tara_Data}
\title{Tara Oceans data}
\format{
\subsection{\code{Tara_Data}}{

Phyloseq-class object with 3 samples.
\describe{
\item{otu_table()}{1701 taxa and 3 samples}
\item{sample_data()}{3 samples by 61 sample variables}
\item{tax_table()}{1701 taxa by 7 taxonomic ranks}
...
}
}
}
\source{
\url{https://www.ebi.ac.uk/metagenomics/analyses/MGYA00133816}, \url{https://www.ebi.ac.uk/metagenomics/analyses/MGYA00133817}, \url{https://www.ebi.ac.uk/metagenomics/analyses/MGYA00133818}
}
\usage{
Tara_Data
}
\description{
A subset (3 samples) of data from Shotgun Sequencing of Tara Oceans DNA samples
corresponding to size fractions for small DNA viruses.
}
\details{
"Tara Expeditions are global scientific voyages that probe morphological and molecular
diversity, evolution and ecology of marine plankton to explore how they are impacted by changes
in the Earth's climate. The first expeditions collected samples of marine plankton containing
viruses, bacteria, archaea, protists and planktonic metazoans living in the photic layer of
the world's oceans. These expeditions, which took place between 2009 and 2013, include Tara
Oceans: a global view, and Tara Oceans Polar Circle, both of which followed the same sampling protocol".
\url{https://www.ebi.ac.uk/services/tara-oceans-data}
}
\keyword{datasets}
