% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weblinks.R
\docType{data}
\name{Weblinks}
\alias{Weblinks}
\title{Weblinks.}
\format{
\subsection{\code{Weblinks}}{

Data frame that contains.
\describe{
\item{degree}{The number of links pointing to the page.}
\item{frequency}{The number of pages that have a given degree. The degree distribution of the network.}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Weblinks
}
\description{
The degree distribution of links between pages in a network built from a 1997 web crawl.
}
\details{
" The dataset contains 'Two AltaVista crawls each with over 200
million pages and 1.5 billion links.'
Broder, A., R. Kumar, F. Maghoul, P. Raghavan, S. Rajagopalan, R. Stata,
A. Tomkins, and J. Wiener, 2000, Computer Networks 33, 309. "
}
\keyword{datasets}
