% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_many.R
\name{BC_many}
\alias{BC_many}
\title{BC_multiple for Many Objects}
\usage{
BC_many(
  object_names,
  column = NULL,
  use_paths = FALSE,
  is_phyloseq = FALSE,
  supress_messages_many = TRUE,
  ...
)
}
\arguments{
\item{object_names}{A vector containing the names of the objects to process.}

\item{column}{Optional. A list that contains vectors that indicate which columns to select for each processed object. Order must be matched with object_names.}

\item{use_paths}{Optional. Set to true when the object_names vector uses paths instead of loaded objects.}

\item{is_phyloseq}{Optional. Set to true if the objects to analyse have the phyloseq class.}

\item{supress_messages_many}{Whether to hide the report progress that indicates how many objects are left to analyse. Also stops the resulting data frame from being printed. Defaults to true.}

\item{...}{Passes arguments to \code{BC_multiple}}
}
\value{
A data frame that contains BC_multiple style information with a column appended to the right that indicates to which object the analysis belongs.
}
\description{
Analyses N objects using \code{BC_multiple} in sequence.
}
\examples{
to_analyze <- c("Weblinks","Billionaires","MOMv3.3")
columns_to_use <-list(2,2,c(7,8))
BC_many(to_analyze,columns_to_use,show_stats=FALSE)
}
