% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_param.R
\name{BC_param}
\alias{BC_param}
\title{Estimate the DGB Distribution Parameters}
\usage{
BC_param(
  df_abundance = NULL,
  column = NULL,
  confidence_interval = 0.95,
  nls = FALSE,
  nls_loop = 1,
  nls_algo = "lm",
  nls_control = list(scale = "more"),
  BC_rank_object = NULL,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{confidence_interval}{Numeric. The confidence interval to calculate for the DGB distribution.}

\item{nls}{Logical. Set to TRUE to use a non-linear least squares fitting method from the 'gslnls' package.}

\item{nls_loop}{Numeric. Set to values higher than 1 to repeat the nls method as many times as needed, reseeding with the last value. This improves the fit with diminishing returns.}

\item{nls_algo}{String. The algorithm to use for the \code{gsl_nls} function.}

\item{nls_control}{List. Provided for the control argument for the \code{gsl_nls} function.}

\item{BC_rank_object}{Optional. A previous object generated by \code{BC_rank}.}

\item{...}{passes arguments to \code{BC_rank}.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data and predicted confidence interval values, the adjusted parameters, the confidence interval of the parameters, the linear model and a summary of the model.
}
\description{
Estimates the parameters of the DGB distribution (also known as the Beta-Cocho distribution) first defined in Mansilla et al. (2007) \doi{10.1016/j.joi.2007.01.001} and further characterized in Martinez-Mekler et al. (2009) \doi{10.1371/journal.pone.0004791} for a given set of data.
\code{BC_param} calculates the log of the data and estimates the abundance from the ranking using a linear model.
The coefficients of the linear model are then scaled for future use.
}
\examples{
BC_param(df_abundance=DGBD::Billionaires, column= 2, confidence_interval=0.99)

BC_param(df_abundance=DGBD::MOMv3.3, column=7, nls=TRUE)

}
