% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_plot.R
\name{BC_plot}
\alias{BC_plot}
\title{Plot DGB Distribution and Observed Data}
\usage{
BC_plot(
  df_abundance = NULL,
  column = NULL,
  BC_model_object = NULL,
  obs = TRUE,
  obs_shape = 16,
  obs_col = "#78a7ff",
  obs_size = 1,
  model = TRUE,
  model_col = "#000000",
  model_width = 0.5,
  model_extra = "MSE",
  confint = TRUE,
  confint_col = "#ed8666",
  confint_width = 1,
  confrange = TRUE,
  confrange_col = "#ffd078",
  gfx_alpha = 0.75,
  gfx_title = "Rank-Abundance Diagram",
  gfx_label = TRUE,
  gfx_label_coords = NULL,
  gfx_xy_trans = c("identity", "log10"),
  gfx_theme = ggplot2::theme_gray(),
  plot_silent = FALSE,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_model_object}{Optional. A previous object generated by \code{BC_model}.}

\item{obs}{Logical. Whether to plot the observed abundance data. Defaults to true.}

\item{obs_shape}{Numerical. The shape of the plotted observed abundance data.}

\item{obs_col}{The color for the observations.}

\item{obs_size}{Numeric. The size for the observations.}

\item{model}{Logical. Whether to show the models predicted data. Defaults to true.}

\item{model_col}{Specify a color for the model.}

\item{model_width}{Numeric. Changes the width of the lines to use for the model.}

\item{model_extra}{String. Has to be one of: "MSE" (Mean Square Error),"S" (Standard error of the Estimate),"R2". Defaults to "MSE".}

\item{confint}{Logical. Whether to add the confidence interval lines. Defaults to true.}

\item{confint_col}{Specify a color for the confidence interval lines.}

\item{confint_width}{Numeric. Changes the width of the confidence interval lines.}

\item{confrange}{Logical. Whether to shade the area in the confidence interval. Defaults to true.}

\item{confrange_col}{Specify a color to use for the confidence interval shading.}

\item{gfx_alpha}{Numeric. Modifies all the graphed objects alpha. Default=0.75.}

\item{gfx_title}{String. Changes the title of the graph.}

\item{gfx_label}{Logical. Whether to show the parameters used and model_extra info.}

\item{gfx_label_coords}{A vector that provides custom x and y values to move the label.}

\item{gfx_xy_trans}{A vector with 2 strings that define the ggplot2 transformations to be applied to the x and y scales. Defaults to c("identity","log10").}

\item{gfx_theme}{Provide a ggplot2 theme function to use. Defaults to theme_gray().}

\item{plot_silent}{Logical. Whether to print to console the output list and plot the graph.}

\item{...}{passes arguments to \code{BC_model}.}
}
\value{
A list with the following elements: The input data frame with added processed ranking data, model data and confidence interval data, the adjusted parameters, the confidence interval of the parameters, the linear model, a summary of the model, a generated function for use with raw numeric data and a ggplot2 object that shows the DGBD distribution and observed data, a model_extra vector with 2 elements, model_extra name and value.
}
\description{
Creates a graphical representation of the DGB Distribution (Mansilla et al. (2007) \doi{10.1016/j.joi.2007.01.001}) model. It supports both linear and nls fits done in \code{BC_param}. Requires a function generated by \code{BC_model}.
}
\examples{
BC_plot(Weblinks, column=2, rank_threshold=4,confint=FALSE,confrange=FALSE,plot_silent=TRUE)

BC_plot(DGBD::hmp_wgs,2,obs=FALSE,plot_silent=TRUE)

BC_plot(Billionaires, column= 2, gfx_xy_trans=c("log10","log10"),plot_silent=TRUE)
}
