% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_random.R
\name{BC_random}
\alias{BC_random}
\title{Generate random points on a DGB Distribution.}
\usage{
BC_random(parameters)
}
\arguments{
\item{parameters}{A numeric vector (A,a,b,max,n).}
}
\value{
A data frame with x,y values
}
\description{
Receives a set of parameters and generates a set of "x" ranking points that
are uniformly random in a linear space and then calculates their expected "y"
or abundance values. Function derived from:
Fontanelli, O., Miramontes, P., Mansilla, R., Cocho, G., & Li, W. (2022).
Beta rank function: A smooth double-Pareto-like distribution. Communications
in Statistics - Theory and Methods, 51(11), 3645–3668.
https://doi.org/10.1080/03610926.2020.1800739
}
\examples{
BC_random(c(25,0.3,0.1,50,4))

plot(BC_random(c(25,0.3,0.1,50,100)))
}
