% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_report.R
\name{BC_report}
\alias{BC_report}
\title{Summarize the Modeled DGB Distribution}
\usage{
BC_report(
  df_abundance = NULL,
  column_report = NULL,
  BC_plot_object = NULL,
  show_plot = TRUE,
  show_stats = TRUE,
  ...
)
}
\arguments{
\item{df_abundance}{A data frame that contains abundance data.}

\item{column_report}{Either a string with the name of the column or the number of the column that stores the abundances in the data frame.}

\item{BC_plot_object}{A previous object generated by \code{BC_plot}.}

\item{show_plot}{Logical. Whether to plot the DGB distribution. Defaults to true.}

\item{show_stats}{Logical. Whether to print the parameters of the DGB distribution and its R2 to the console. Defaults to true.}

\item{...}{passes arguments to \code{BC_plot}.}
}
\value{
A list with the following elements: a matrix that contains the parameters used and their confidence intervals, the R2 and the DGBD plot.
}
\description{
Summarize the Modeled DGB Distribution
}
\examples{
BC_report(hmp_wgs,2)

BC_report(hmp_wgs,2,show_plot=FALSE,confidence_interval=0.99)

BC_report(Tara_Data,2,is_phyloseq=TRUE)

}
