% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weblinks.R
\docType{data}
\name{Weblinks}
\alias{Weblinks}
\title{Weblinks.}
\format{
\subsection{\code{Weblinks}}{

Data frame that contains.
\describe{
\item{degree}{The number of links pointing to a page}
\item{frequency}{The number of pages that have a given degree}
...
}
}
}
\source{
\url{https://github.com/robertoalvarezm/DataNonPowerLaws}
}
\usage{
Weblinks
}
\description{
The number of links to websites found in a 1997 web crawl of about 200 million web pages is represented as a simple histogram.
}
\details{
""
}
\keyword{datasets}
